finish 2 task
Wrote a script where was used an array of image objects to create <img> elements nested inside <li> elements. The HTML elements was generated either by using document.createElement() along with the append() method or by employing template strings with insertAdjacentHTML(). All of the gallery items were added to the DOM in a single operation to ensure efficiency.