finish task 3
A script that, while typing text in the input#name-input (input event), inserts its current value into the span#name-output as the name for the greeting. To trim the input value to remove leading and trailing spaces. If the input is empty or contains only spaces, the span should display the string "Anonymous" instead of the name.