finish ttsk 7
In this code, the user can enter their name, age, and choose an avatar color using the form. After clicking the "Add" button, the user's information is displayed as a card in the list below.

Each card contains:

The user's name and age.
A circular avatar with the selected color.
A "Delete" button to remove the card.
If the user's age is under 18, the text will be displayed in red. If the age is 18 or older, the text will be displayed in green.

The form also validates the input fields, showing an alert message if the fields are empty or the age is less than 0.